hamza

+++ src/Token/index.js
@@ -1,5 +1,9 @@
 import baseUrl from "../config.json";
+import { useNavigate } from 'react-router-dom'
+
+
 export const getToken = async () => {
+    // const navigate = useNavigate()
     try {
         const response = await fetch(
             `${baseUrl.baseUrl}/auth/Authorization`,
@@ -12,10 +16,10 @@ export const getToken = async () => {
             }
         );
         const res = await response.json();
-        if (response.status === 200) {
+        if (response.status == 200) {
             return res
         } else {
-            if (res?.messsage == "timeout error" || res?.messsage == "unauthorized") {
+            if (res?.messsage === "timeout error" || res?.messsage == "unauthorized") {
                 const response = await fetch(
                 `${baseUrl.baseUrl}/auth/Authorization`,{
                     method: "GET",
@@ -28,11 +32,15 @@ export const getToken = async () => {
                     const res = await response.json();
                     localStorage.setItem("refresh", res?.refresh_token)
                     localStorage.setItem("access_token", res?.access_token)
+                    console.log("res",res)
                     return res
                 }
                 else{
-                    window.location.href = '/'
-                    localStorage.clear()
+                    if (window.location.pathname !== '/') {
+                        window.location.href = '/';
+                        localStorage.clear()
+                        return res
+                    }
                 }
             }
         }
